requirements.txt
pymupdf
nltk
sentence-transformers
faiss-cpu